requirements.txt
ttkbootstrap==1.10.1
pillow==10.3.0
